requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv==1.0.0

# Optional but recommended for better terminal output
colorama==0.4.6
tqdm==4.65.0

# Development dependencies
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0 